Implement a dbus library
Initial implementation of a dbus library.

Current usage is a little clunky. Users of this library need to
* Use sphtud themselves
* Generate a wrapper interface with the provided "generate" binary
* Add sphtud and sphdbus modules as imports to the generated module
* Likely override the sphtud module import in the sphdbus module itself

In it's current form, we are good enough to do a couple property
gets/sets on spotify, and a couple function calls. We did a little bit
of testing on the login1 freedesktop interface during development as
well, but that is not explicitly tested in our final form

Lots of stuff probably won't work. We haven't been very thorough about
implementing serializers/deserializers for all types. Anything that
passes fds will be fully skipped in API generation, but lots of other
stuff might just not work when we actually try them. We can fix these
later

This is mostly based off the dbus spec at
https://dbus.freedesktop.org/doc/dbus-specification.html as well as
stracing qdbus interactions with spotify